+++ src/scene/export.ts
@@ -1,21 +1,13 @@
 import rough from "roughjs/bin/rough";
-import oc from "open-color";
-import { newTextElement } from "../element";
 import { NonDeletedExcalidrawElement } from "../element/types";
 import { getCommonBounds } from "../element/bounds";
 import { renderScene, renderSceneToSvg } from "../renderer/renderScene";
 import { distance, SVG_NS } from "../utils";
 import { AppState } from "../types";
-import { t } from "../i18n";
-import {
-  DEFAULT_FONT_FAMILY,
-  DEFAULT_VERTICAL_ALIGN,
-  THEME_FILTER,
-} from "../constants";
+import { THEME_FILTER } from "../constants";
 import { getDefaultAppState } from "../appState";
 
 export const SVG_EXPORT_TAG = `<!-- svg-source:excalidraw -->`;
-const WATERMARK_HEIGHT = 16;
 
 export const exportToCanvas = (
   elements: readonly NonDeletedExcalidrawElement[],
@@ -25,13 +17,11 @@ export const exportToCanvas = (
     exportPadding = 10,
     viewBackgroundColor,
     scale = 1,
-    shouldAddWatermark,
   }: {
     exportBackground: boolean;
     exportPadding?: number;
     scale?: number;
     viewBackgroundColor: string;
-    shouldAddWatermark: boolean;
   },
   createCanvas: (
     width: number,
@@ -43,13 +33,7 @@ export const exportToCanvas = (
     return { canvas: tempCanvas, scale };
   },
 ) => {
-  const sceneElements = getElementsAndWatermark(elements, shouldAddWatermark);
-
-  const [minX, minY, width, height] = getCanvasSize(
-    sceneElements,
-    exportPadding,
-    shouldAddWatermark,
-  );
+  const [minX, minY, width, height] = getCanvasSize(elements, exportPadding);
 
   const { canvas: tempCanvas, scale: newScale = scale } = createCanvas(
     width,
@@ -57,7 +41,7 @@ export const exportToCanvas = (
   );
 
   renderScene(
-    sceneElements,
+    elements,
     appState,
     null,
     newScale,
@@ -94,7 +78,6 @@ export const exportToSvg = (
     viewBackgroundColor,
     exportWithDarkMode,
     scale = 1,
-    shouldAddWatermark,
     metadata = "",
   }: {
     exportBackground: boolean;
@@ -102,17 +85,10 @@ export const exportToSvg = (
     scale?: number;
     viewBackgroundColor: string;
     exportWithDarkMode?: boolean;
-    shouldAddWatermark: boolean;
     metadata?: string;
   },
 ): SVGSVGElement => {
-  const sceneElements = getElementsAndWatermark(elements, shouldAddWatermark);
-
-  const [minX, minY, width, height] = getCanvasSize(
-    sceneElements,
-    exportPadding,
-    shouldAddWatermark,
-  );
+  const [minX, minY, width, height] = getCanvasSize(elements, exportPadding);
 
   // initialze SVG root
   const svgRoot = document.createElementNS(SVG_NS, "svg");
@@ -154,7 +130,7 @@ export const exportToSvg = (
   }
 
   const rsvg = rough.svg(svgRoot);
-  renderSceneToSvg(sceneElements, rsvg, svgRoot, {
+  renderSceneToSvg(elements, rsvg, svgRoot, {
     offsetX: -minX + exportPadding,
     offsetY: -minY + exportPadding,
   });
@@ -162,52 +138,14 @@ export const exportToSvg = (
   return svgRoot;
 };
 
-const getElementsAndWatermark = (
-  elements: readonly NonDeletedExcalidrawElement[],
-  shouldAddWatermark: boolean,
-): readonly NonDeletedExcalidrawElement[] => {
-  let _elements = [...elements];
-
-  if (shouldAddWatermark) {
-    const [, , maxX, maxY] = getCommonBounds(elements);
-    _elements = [..._elements, getWatermarkElement(maxX, maxY)];
-  }
-
-  return _elements;
-};
-
-const getWatermarkElement = (maxX: number, maxY: number) => {
-  return newTextElement({
-    text: t("labels.madeWithExcalidraw"),
-    fontSize: WATERMARK_HEIGHT,
-    fontFamily: DEFAULT_FONT_FAMILY,
-    textAlign: "right",
-    verticalAlign: DEFAULT_VERTICAL_ALIGN,
-    x: maxX,
-    y: maxY + WATERMARK_HEIGHT,
-    strokeColor: oc.gray[5],
-    backgroundColor: "transparent",
-    fillStyle: "hachure",
-    strokeWidth: 1,
-    strokeStyle: "solid",
-    roughness: 1,
-    opacity: 100,
-    strokeSharpness: "sharp",
-  });
-};
-
 // calculate smallest area to fit the contents in
 const getCanvasSize = (
   elements: readonly NonDeletedExcalidrawElement[],
   exportPadding: number,
-  shouldAddWatermark: boolean,
 ): [number, number, number, number] => {
   const [minX, minY, maxX, maxY] = getCommonBounds(elements);
   const width = distance(minX, maxX) + exportPadding * 2;
-  const height =
-    distance(minY, maxY) +
-    exportPadding +
-    (shouldAddWatermark ? 0 : exportPadding);
+  const height = distance(minY, maxY) + exportPadding + exportPadding;
 
   return [minX, minY, width, height];
 };
@@ -215,15 +153,11 @@ const getCanvasSize = (
 export const getExportSize = (
   elements: readonly NonDeletedExcalidrawElement[],
   exportPadding: number,
-  shouldAddWatermark: boolean,
   scale: number,
 ): [number, number] => {
-  const sceneElements = getElementsAndWatermark(elements, shouldAddWatermark);
-
   const [, , width, height] = getCanvasSize(
-    sceneElements,
+    elements,
     exportPadding,
-    shouldAddWatermark,
   ).map((dimension) => Math.trunc(dimension * scale));
 
   return [width, height];